Include game_state header and move_generator
