+++ CMakeLists.txt
@@ -109,7 +109,8 @@ add_library(libblaengine SHARED
     source/blaengine.cpp
     source/communication/uci.cpp
     source/calculation/calculator.cpp
-    source/calculation/game_state_controller.cpp)
+    source/calculation/game_state_controller.cpp
+    source/calculation/move_generator.cpp)
 
 target_link_libraries(libblaengine
     ${GTEST_LIBRARIES}
